1. rename proxy.go, proxy_test.go into client.go and client_test.go 2. rewrite DryRunRequest to use the request context
